+++ log/pretty.go
@@ -15,6 +15,18 @@ func PrettyString(i interface{}) string {
 	return spew.Sdump(i)
 }
 
+func ColorSprint(i interface{}) string {
+	return pp.Sprint(i)
+}
+
+func ColorSprintln(i interface{}) string {
+	return pp.Sprintln(i)
+}
+
+func ColorSprintf(fmt string, args ...interface{}) string {
+	return pp.Sprintf(fmt, args...)
+}
+
 func ColorPrint(i interface{}) {
 	pp.Print(i)
 }